[OT-104] [REFACTOR]: 업로드 DTO 패키지 경로 정리 및 import 수정
