+++ apps/api-admin/src/main/java/com/ott/api_admin/series/controller/BackOfficeSeriesController.java
@@ -1,14 +1,24 @@
 package com.ott.api_admin.series.controller;
 
+import com.ott.api_admin.series.dto.request.SeriesUploadRequest;
 import com.ott.api_admin.series.dto.response.SeriesDetailResponse;
 import com.ott.api_admin.series.dto.response.SeriesListResponse;
 import com.ott.api_admin.series.dto.response.SeriesTitleListResponse;
+import com.ott.api_admin.series.dto.response.SeriesUploadResponse;
 import com.ott.api_admin.series.service.BackOfficeSeriesService;
 import com.ott.common.web.response.PageResponse;
 import com.ott.common.web.response.SuccessResponse;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/back-office/admin/series")
@@ -47,5 +57,14 @@ public class BackOfficeSeriesController implements BackOfficeSeriesApi {
         return ResponseEntity.ok(
                 SuccessResponse.of(backOfficeSeriesService.getSeriesDetail(mediaId))
         );
+    }
+
+    @Override
+    @PostMapping("/upload")
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<SuccessResponse<SeriesUploadResponse>> createSeriesUpload(
+            @Valid @RequestBody SeriesUploadRequest request
+    ) {
+        return ResponseEntity.ok(SuccessResponse.of(backOfficeSeriesService.createSeriesUpload(request)));
     }
 }
